Moving license installer proxy to its own file
