+++ dev/Deployment/Deployment.vcxitems
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup Label="Globals">
     <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
@@ -30,6 +30,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)Deployer.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)PackageUtilities.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)PackageRegistrar.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)LicenseInstallerProxy.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)DeploymentTraceLogging.cpp" />
